Refactor parallelization
